// origin/src/youtube/types/YTCFG.ts
export interface YTCFG {
    CLIENT_CANARY_STATE: string
    DEVICE: string
    DISABLE_YT_IMG_DELAY_LOADING: boolean
    ELEMENT_POOL_DEFAULT_CAP: number
    EVENT_ID: string
    EXPERIMENT_FLAGS: ExperimentFlags
    GAPI_HINT_PARAMS: string
    GAPI_HOST: string
    GAPI_LOCALE: string
    GL: string
    GOOGLE_FEEDBACK_PRODUCT_ID: string
    GOOGLE_FEEDBACK_PRODUCT_DATA: GoogleFeedbackProductData
    HL: string
    HTML_DIR: string
    HTML_LANG: string
    ID_TOKEN: string
    INNERTUBE_API_KEY: string
    INNERTUBE_API_VERSION: string
    INNERTUBE_CLIENT_NAME: string
    INNERTUBE_CLIENT_VERSION: string
    INNERTUBE_CONTEXT: InnertubeContext
    INNERTUBE_CONTEXT_CLIENT_NAME: number
    INNERTUBE_CONTEXT_CLIENT_VERSION: string
    INNERTUBE_CONTEXT_GL: string
    INNERTUBE_CONTEXT_HL: string
    LATEST_ECATCHER_SERVICE_TRACKING_PARAMS: LatestEcatcherServiceTrackingParams
    LOGGED_IN: boolean
    PAGE_BUILD_LABEL: string
    PAGE_CL: number
    scheduler: Scheduler
    SERVER_NAME: string
    SESSION_INDEX: string
    SIGNIN_URL: string
    VISITOR_DATA: string
    WEB_PLAYER_CONTEXT_CONFIGS: WebPlayerContextConfigs
    XSRF_FIELD_NAME: string
    XSRF_TOKEN: string
    YPC_MB_URL: string
    YTR_FAMILY_CREATION_URL: string
    SERVER_VERSION: string
    REUSE_COMPONENTS: boolean
    STAMPER_STABLE_LIST: boolean
    DATASYNC_ID: string
    SERIALIZED_CLIENT_CONFIG_DATA: string
    LOCATION_PLAYABILITY_TOKEN: string
    LIVE_CHAT_BASE_TANGO_CONFIG: LiveChatBaseTangoConfig
    FEXP_EXPERIMENTS: number[]
    LIVE_CHAT_SEND_MESSAGE_ACTION: string
    ROOT_VE_TYPE: number
    CLIENT_PROTOCOL: string
    CLIENT_TRANSPORT: string
    TIME_CREATED_MS: number
    BUTTON_REWORK: boolean
    LOGIN_INFO: string
    VALID_SESSION_TEMPDATA_DOMAINS: string[]
    WORKER_PERFORMANCE_URL: WorkerPerformanceUrl
    RAW_COLD_CONFIG_GROUP: RawColdConfigGroup
    RAW_HOT_CONFIG_GROUP: RawHotConfigGroup
    SERIALIZED_HOT_HASH_DATA: string
    SERIALIZED_COLD_HASH_DATA: string
    PERSIST_IDENTITY_IFRAME_URL: PersistIdentityIframeUrl
    WORKER_SERIALIZATION_URL: WorkerSerializationUrl
    LOG_STICKINESS: boolean
    USER_SESSION_ID: string
    VISIBILITY_TIME_BETWEEN_JOBS_MS: number
    START_IN_THEATER_MODE: boolean
    START_IN_FULL_WINDOW_MODE: boolean
    SERVICE_WORKER_PROMPT_NOTIFICATIONS: boolean
    SBOX_LABELS: SboxLabels
    ONE_PICK_URL: string
    NO_EMPTY_DATA_IMG: boolean
    MENTIONS_EDU_HELP_LINK: string
    DEFERRED_DETACH: boolean
    RECAPTCHA_V3_SITEKEY: string
    PLAYER_JS_URL: string
    PLAYER_CSS_URL: string
    LINK_GAL_DOMAIN: string
    LINK_OIS_DOMAIN: string
    IS_TABLET: boolean
    LINK_API_KEY: string
    DISABLE_WARM_LOADS: boolean
    ZWIEBACK_PING_URLS: string[]
    VOZ_API_KEY: string
    STS: number
    SBOX_SETTINGS: SboxSettings
    SBOX_JS_URL: string
}
// ...
export interface GoogleFeedbackProductData {
    polymer: string
    polymer2: string
    accept_language: string
}

export interface InnertubeContext {
    client: Client
    user: User
    request: Request
    clickTracking: ClickTracking
}

export interface Client {
    hl: string
    gl: string
    remoteHost: string
    deviceMake: string
    deviceModel: string
    visitorData: string
    userAgent: string
    clientName: string
    clientVersion: string
    osName: string
    osVersion: string
    originalUrl: string
    platform: string
    clientFormFactor: string
    configInfo: ConfigInfo
    userInterfaceTheme: string
    timeZone: string
    browserName: string
    browserVersion: string
    locationPlayabilityToken: string
    acceptHeader: string
    deviceExperimentId: string
}

export interface ConfigInfo {
    appInstallData: string
}

export interface User {
    lockedSafetyMode: boolean
}

export interface Request {
    useSsl: boolean
    consistencyTokenJars: ConsistencyTokenJar[]
}

export interface ConsistencyTokenJar {
    encryptedTokenJarContents: string
}

export interface ClickTracking {
    clickTrackingParams: string
}

export interface LatestEcatcherServiceTrackingParams {
    "client.name": string
}

export interface Scheduler {
    useRaf: boolean
    timeout: number
}

export interface WebPlayerContextConfigs {
    WEB_PLAYER_CONTEXT_CONFIG_ID_KEVLAR_WATCH: WebPlayerContextConfigIdKevlarWatch
    WEB_PLAYER_CONTEXT_CONFIG_ID_KEVLAR_CHANNEL_TRAILER: WebPlayerContextConfigIdKevlarChannelTrailer
    WEB_PLAYER_CONTEXT_CONFIG_ID_KEVLAR_PLAYLIST_OVERVIEW: WebPlayerContextConfigIdKevlarPlaylistOverview
    WEB_PLAYER_CONTEXT_CONFIG_ID_KEVLAR_VERTICAL_LANDING_PAGE_PROMO: WebPlayerContextConfigIdKevlarVerticalLandingPagePromo
    WEB_PLAYER_CONTEXT_CONFIG_ID_KEVLAR_SHORTS: WebPlayerContextConfigIdKevlarShorts
    WEB_PLAYER_CONTEXT_CONFIG_ID_KEVLAR_SPONSORSHIPS_OFFER: WebPlayerContextConfigIdKevlarSponsorshipsOffer
    WEB_PLAYER_CONTEXT_CONFIG_ID_KEVLAR_INLINE_PREVIEW: WebPlayerContextConfigIdKevlarInlinePreview
    WEB_PLAYER_CONTEXT_CONFIG_ID_HANDLES_CLAIMING: WebPlayerContextConfigIdHandlesClaiming
    WEB_PLAYER_CONTEXT_CONFIG_ID_KEVLAR_SFV_AUDIO_ITEM: WebPlayerContextConfigIdKevlarSfvAudioItem
    WEB_PLAYER_CONTEXT_CONFIG_ID_KEVLAR_MASTHEAD_AD: WebPlayerContextConfigIdKevlarMastheadAd
}

export interface WebPlayerContextConfigIdKevlarWatch {
    transparentBackground: boolean
    showMiniplayerButton: boolean
    externalFullscreen: boolean
    showMiniplayerUiWhenMinimized: boolean
    rootElementId: string
    jsUrl: string
    cssUrl: string
    contextId: string
    eventLabel: string
    contentRegion: string
    hl: string
    hostLanguage: string
    playerStyle: string
    innertubeApiKey: string
    innertubeApiVersion: string
    innertubeContextClientVersion: string
    device: Device
    serializedExperimentIds: string
    serializedExperimentFlags: string
    cspNonce: string
    canaryState: string
    enableCsiLogging: boolean
    csiPageType: string
    authorizedUserIndex: number
    datasyncId: string
    allowWoffleManagement: boolean
    cinematicSettingsAvailable: boolean
    canaryStage: string
}

export interface Device {
    brand: string
    model: string
    browser: string
    browserVersion: string
    os: string
    osVersion: string
    platform: string
    interfaceName: string
    interfaceVersion: string
}

export interface WebPlayerContextConfigIdKevlarChannelTrailer {
    rootElementId: string
    jsUrl: string
    cssUrl: string
    contextId: string
    eventLabel: string
    contentRegion: string
    hl: string
    hostLanguage: string
    playerStyle: string
    innertubeApiKey: string
    innertubeApiVersion: string
    innertubeContextClientVersion: string
    device: Device2
    serializedExperimentIds: string
    serializedExperimentFlags: string
    cspNonce: string
    canaryState: string
    enableCsiLogging: boolean
    csiPageType: string
    authorizedUserIndex: number
    datasyncId: string
    canaryStage: string
}

export interface Device2 {
    brand: string
    model: string
    browser: string
    browserVersion: string
    os: string
    osVersion: string
    platform: string
    interfaceName: string
    interfaceVersion: string
}

export interface WebPlayerContextConfigIdKevlarPlaylistOverview {
    rootElementId: string
    jsUrl: string
    cssUrl: string
    contextId: string
    eventLabel: string
    contentRegion: string
    hl: string
    hostLanguage: string
    playerStyle: string
    innertubeApiKey: string
    innertubeApiVersion: string
    innertubeContextClientVersion: string
    device: Device3
    serializedExperimentIds: string
    serializedExperimentFlags: string
    disableSharing: boolean
    hideInfo: boolean
    disableWatchLater: boolean
    cspNonce: string
    canaryState: string
    enableCsiLogging: boolean
    csiPageType: string
    authorizedUserIndex: number
    datasyncId: string
    canaryStage: string
}

export interface Device3 {
    brand: string
    model: string
    browser: string
    browserVersion: string
    os: string
    osVersion: string
    platform: string
    interfaceName: string
    interfaceVersion: string
}

export interface WebPlayerContextConfigIdKevlarVerticalLandingPagePromo {
    rootElementId: string
    jsUrl: string
    cssUrl: string
    contextId: string
    eventLabel: string
    contentRegion: string
    hl: string
    hostLanguage: string
    playerStyle: string
    innertubeApiKey: string
    innertubeApiVersion: string
    innertubeContextClientVersion: string
    controlsType: number
    disableRelatedVideos: boolean
    annotationsLoadPolicy: number
    device: Device4
    serializedExperimentIds: string
    serializedExperimentFlags: string
    hideInfo: boolean
    startMuted: boolean
    enableMutedAutoplay: boolean
    cspNonce: string
    canaryState: string
    enableCsiLogging: boolean
    csiPageType: string
    authorizedUserIndex: number
    datasyncId: string
    canaryStage: string
}

export interface Device4 {
    brand: string
    model: string
    browser: string
    browserVersion: string
    os: string
    osVersion: string
    platform: string
    interfaceName: string
    interfaceVersion: string
}

export interface WebPlayerContextConfigIdKevlarShorts {
    rootElementId: string
    jsUrl: string
    cssUrl: string
    contextId: string
    eventLabel: string
    contentRegion: string
    hl: string
    hostLanguage: string
    playerStyle: string
    innertubeApiKey: string
    innertubeApiVersion: string
    innertubeContextClientVersion: string
    controlsType: number
    disableKeyboardControls: boolean
    disableRelatedVideos: boolean
    annotationsLoadPolicy: number
    device: Device5
    serializedExperimentIds: string
    serializedExperimentFlags: string
    hideInfo: boolean
    disableFullscreen: boolean
    cspNonce: string
    canaryState: string
    enableCsiLogging: boolean
    authorizedUserIndex: number
    datasyncId: string
    storeUserVolume: boolean
    disableSeek: boolean
    disablePaidContentOverlay: boolean
    preferGapless: boolean
    canaryStage: string
}

export interface Device5 {
    brand: string
    model: string
    browser: string
    browserVersion: string
    os: string
    osVersion: string
    platform: string
    interfaceName: string
    interfaceVersion: string
}

export interface WebPlayerContextConfigIdKevlarSponsorshipsOffer {
    rootElementId: string
    jsUrl: string
    cssUrl: string
    contextId: string
    eventLabel: string
    contentRegion: string
    hl: string
    hostLanguage: string
    playerStyle: string
    innertubeApiKey: string
    innertubeApiVersion: string
    innertubeContextClientVersion: string
    disableRelatedVideos: boolean
    annotationsLoadPolicy: number
    device: Device6
    serializedExperimentIds: string
    serializedExperimentFlags: string
    disableFullscreen: boolean
    cspNonce: string
    canaryState: string
    authorizedUserIndex: number
    datasyncId: string
    canaryStage: string
}

export interface Device6 {
    brand: string
    model: string
    browser: string
    browserVersion: string
    os: string
    osVersion: string
    platform: string
    interfaceName: string
    interfaceVersion: string
}

export interface WebPlayerContextConfigIdKevlarInlinePreview {
    rootElementId: string
    jsUrl: string
    cssUrl: string
    contextId: string
    eventLabel: string
    contentRegion: string
    hl: string
    hostLanguage: string
    playerStyle: string
    innertubeApiKey: string
    innertubeApiVersion: string
    innertubeContextClientVersion: string
    controlsType: number
    disableKeyboardControls: boolean
    device: Device7
    serializedExperimentIds: string
    serializedExperimentFlags: string
    disableFullscreen: boolean
    cspNonce: string
    canaryState: string
    enableCsiLogging: boolean
    csiPageType: string
    authorizedUserIndex: number
    disableMdxCast: boolean
    datasyncId: string
    disablePaidContentOverlay: boolean
    canaryStage: string
}
// ...
export interface LiveChatBaseTangoConfig {
    apiKey: string
    channelUri: string
    clientName: string
    requiresAuthToken: boolean
    senderUri: string
    useNewTango: boolean
}

export interface WorkerPerformanceUrl {
    privateDoNotAccessOrElseTrustedResourceUrlWrappedValue: string
}

export interface RawColdConfigGroup {
    configData: string
    mainAppColdConfig: MainAppColdConfig
}
// ...
export interface RawHotConfigGroup {
    mainAppHotConfig: MainAppHotConfig
    loggingHotConfig: LoggingHotConfig
}
// ...
export interface LoggingHotConfig {
    eventLoggingConfig: EventLoggingConfig
    csiConfig: CsiConfig
}
// ...
export interface PersistIdentityIframeUrl {
    privateDoNotAccessOrElseTrustedResourceUrlWrappedValue: string
}

export interface WorkerSerializationUrl {
    privateDoNotAccessOrElseTrustedResourceUrlWrappedValue: string
}

export interface SboxLabels {
    SUGGESTION_DISMISS_LABEL: string
    SUGGESTION_DISMISSED_LABEL: string
}

export interface SboxSettings {
    HAS_ON_SCREEN_KEYBOARD: boolean
    IS_FUSION: boolean
    IS_POLYMER: boolean
    PSUGGEST_TOKEN: string
    REQUEST_DOMAIN: string
    REQUEST_LANGUAGE: string
    SEND_VISITOR_DATA: boolean
    SEARCHBOX_BEHAVIOR_EXPERIMENT: string
    SEARCHBOX_ENABLE_REFINEMENT_SUGGEST: boolean
    SEARCHBOX_TAP_TARGET_EXPERIMENT: number
    SEARCHBOX_ZERO_TYPING_SUGGEST_USE_REGULAR_SUGGEST: string
    SUGG_EXP_ID: string
    SEARCHBOX_HOST_OVERRIDE: string
    HIDE_REMOVE_LINK: boolean
}